Add Composables function to get team and display it in teams.vue
This commit adds a new Composables function that allows the user to fetch a team object from the API. This team object is then displayed in the teams.vue file using the newly added Composables function. The function fetches all necessary data from the API and makes it available for easy display in the template.